Updating tests, import paths
