+++ src/encoded/static/components/__tests__/workflow-graph-test.js
@@ -11,6 +11,8 @@ jest.autoMockOff();
 jest.dontMock('react');
 jest.dontMock('underscore');
 
+jest.mock('./../util/navigate');
+
 describe('Testing Workflow Graph', function() {
     var context, schemas, testWorkflowInstance, sinon, server;
 
@@ -38,7 +40,8 @@ describe('Testing Workflow Graph', function() {
         var viewProps = {
             schemas, context,
             'onNodeClick' : null,
-            'windowWidth' : 1200
+            'windowWidth' : 1200,
+            'href' : "https://data.4dnucleome.org/workflow-runs-sbg/0fe19e8e-c565-4c0c-8058-effe5d81f53b/"
         };
 
         testWorkflowInstance = TestUtils.renderIntoDocument(<WorkflowRunView {...viewProps} />);
@@ -230,7 +233,8 @@ describe('Find nodes from other columns', function() {
         var viewProps = {
             schemas, context,
             'onNodeClick' : null,
-            'windowWidth' : 1200
+            'windowWidth' : 1200,
+            'href' : "https://data.4dnucleome.org/workflow-runs-sbg/0fe19e8e-c565-4c0c-8058-effe5d81f53b/"
         };
 
         testWorkflowInstance = TestUtils.renderIntoDocument(<WorkflowRunView {...viewProps} />);
